+++ src/views/HomeView.vue
@@ -1,7 +1,9 @@
 <template>
     <HelloWorld />
+    <CompanyRegistrationPopUp />
 </template>
 
 <script lang="ts" setup>
 import HelloWorld from "@/components/HelloWorld.vue";
+import CompanyRegistrationPopUp from "@/components/CompanyRegistrationPopUp.vue";
 </script>
